+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0
+
+- feat: introduction of `PureNotifier`, a `Notifier` without access to `ref`
+- **BREAKING**: add `ref` as parameter to every provider
+- **BREAKING**: change `ref.notify` to `ref.notifier`
+
 ## 0.1.1
 
 - docs: update README.md